Added so that when the game stars, the window is maximized, and fixed a bug when using the window maximize function.
Some incapsulation of a b2Body so its "easier" to create multiple objects and drawing them.

A start of the ground-class have been created.

Some diffs in the project settnings have been made so release-mode works (the solution box2d in external must be compiled in debug and release so both kinds of libs are created..)